+++ packages/xo-web/src/xo-app/vm/tab-network.js
@@ -3,6 +3,8 @@ import ActionButton from 'action-button'
 import ActionRowButton from 'action-row-button'
 import BaseComponent from 'base-component'
 import copy from 'copy-to-clipboard'
+import decorate from 'apply-decorators'
+import getEventValue from 'get-event-value'
 import Icon, { StackedIcons } from 'icon'
 import PropTypes from 'prop-types'
 import React, { Component } from 'react'
@@ -11,11 +13,14 @@ import StateButton from 'state-button'
 import SingleLineRow from 'single-line-row'
 import TabButton from 'tab-button'
 import Tooltip from 'tooltip'
-import { confirm } from 'modal'
-import { isIp, isIpV4 } from 'ip-utils'
+import { confirm, form } from 'modal'
 import { Container, Row, Col } from 'grid'
+import { error } from 'notification'
+import { get } from '@xen-orchestra/defined'
 import { injectIntl } from 'react-intl'
+import { isIp, isIpV4 } from 'ip-utils'
 import { Number, Text, XoSelect } from 'editable'
+import { provideState, injectState } from 'reaclette'
 import {
   addSubscriptions,
   connectStore,
@@ -59,6 +64,7 @@ import {
   deleteVif,
   deleteVifs,
   disconnectVif,
+  getLockingModeValues,
   isVmRunning,
   setVif,
   subscribeIpPools,
@@ -267,6 +273,12 @@ class VifAllowedIps extends BaseComponent {
 }
 
 class VifStatus extends BaseComponent {
+  componentDidMount() {
+    getLockingModeValues().then(lockingModeValues =>
+      this.setState({ lockingModeValues })
+    )
+  }
+
   _getIps = createSelector(
     () => this.props.vif.allowedIpv4Addresses || EMPTY_ARRAY,
     () => this.props.vif.allowedIpv6Addresses || EMPTY_ARRAY,
@@ -281,7 +293,7 @@ class VifStatus extends BaseComponent {
 
     if (lockingMode === 'disabled') {
       return (
-        <Tooltip content={_('vifDisabledNetwork')}>
+        <Tooltip content={_('vifLockingModeDisabled')}>
           <Icon icon='vif-disable' />
         </Tooltip>
       )
@@ -289,7 +301,7 @@ class VifStatus extends BaseComponent {
 
     if (lockingMode === 'unlocked') {
       return (
-        <Tooltip content={_('vifUnLockedNetwork')}>
+        <Tooltip content={_('vifLockingModeUnlocked')}>
           <Icon icon='unlock' />
         </Tooltip>
       )
@@ -297,7 +309,7 @@ class VifStatus extends BaseComponent {
 
     if (lockingMode === 'locked') {
       return (
-        <Tooltip content={_('vifLockedNetwork')}>
+        <Tooltip content={_('vifLockingModeLocked')}>
           <Icon icon='lock' />
         </Tooltip>
       )
@@ -313,7 +325,7 @@ class VifStatus extends BaseComponent {
     }
     if (network.defaultIsLocked) {
       return (
-        <Tooltip content={_('vifLockedNetwork')}>
+        <Tooltip content={_('networkDefaultLockingModeDisabled')}>
           <StackedIcons
             icons={[
               { icon: 'vif-disable', size: 1 },
@@ -324,7 +336,7 @@ class VifStatus extends BaseComponent {
       )
     }
     return (
-      <Tooltip content={_('vifUnLockedNetwork')}>
+      <Tooltip content={_('networkDefaultLockingModeUnlocked')}>
         <StackedIcons
           icons={[
             { icon: 'unlock', size: 1 },
@@ -335,8 +347,14 @@ class VifStatus extends BaseComponent {
     )
   }
 
+  _onChangeVif = event =>
+    setVif(this.props.vif, { lockingMode: getEventValue(event) }).catch(err =>
+      error(_('editVifLockingMode'), err.message || String(err))
+    )
+
   render() {
     const { vif } = this.props
+    const { isLockingModeEdition } = this.state
 
     return (
       <div>
@@ -350,10 +368,80 @@ class VifStatus extends BaseComponent {
           handlerParam={vif}
           state={vif.attached}
         />{' '}
-        {this._getNetworkStatus()}
+        {this._getNetworkStatus()}{' '}
+        {isLockingModeEdition ? (
+          <select
+            className='form-control'
+            onBlur={this.toggleState('isLockingModeEdition')}
+            onChange={this._onChangeVif}
+            value={vif.lockingMode}
+          >
+            {map(this.state.lockingModeValues, lockingMode => (
+              <option key={lockingMode} value={lockingMode}>
+                {lockingMode}
+              </option>
+            ))}
+          </select>
+        ) : (
+          <ActionButton
+            btnStyle='primary'
+            icon='edit'
+            handler={this.toggleState('isLockingModeEdition')}
+            size='small'
+            tooltip={_('editVifLockingMode')}
+          />
+        )}
       </div>
     )
   }
+}
+
+const AdvancedSettingsModal = decorate([
+  provideState({
+    effects: {
+      toggleTxChecksumming() {
+        const { onChange, value } = this.props
+        onChange({
+          ...value,
+          txChecksumming: !value.txChecksumming,
+        })
+      },
+    },
+  }),
+  injectState,
+  ({ effects, state, value }) => (
+    <Container>
+      <SingleLineRow>
+        <Col mediumSize={6}>
+          <strong>{_('txChecksumming')}</strong>
+        </Col>
+        <Col mediumSize={6}>
+          <StateButton
+            className='pull-right'
+            disabledLabel={_('stateDisabled')}
+            enabledLabel={_('stateEnabled')}
+            handler={effects.toggleTxChecksumming}
+            state={value.txChecksumming}
+          />
+        </Col>
+      </SingleLineRow>
+    </Container>
+  ),
+])
+
+const openAdvancedSettingsModal = async vif => {
+  const { txChecksumming } = await form({
+    defaultValue: {
+      txChecksumming: vif.txChecksumming,
+    },
+    header: (
+      <div>
+        <Icon icon='settings' /> {_('advancedSettings')}
+      </div>
+    ),
+    render: props => <AdvancedSettingsModal {...props} />,
+  })
+  await setVif(vif, { txChecksumming })
 }
 
 // -----------------------------------------------------------------------------
@@ -678,6 +766,12 @@ const INDIVIDUAL_ACTIONS = [
     handler: vif => copy(vif.uuid),
     icon: 'clipboard',
     label: vif => _('copyUuid', { uuid: vif.uuid }),
+  },
+  {
+    handler: openAdvancedSettingsModal,
+    icon: 'settings',
+    label: _('advancedSettings'),
+    level: 'primary',
   },
   {
     disabled: vif => vif.attached,